# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora Enterprise Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dealers:
    get:
      operationId: listDealers
      tags: [dealers]
      summary: List dealers
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Dealer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealerList"
    post:
      operationId: createDealer
      tags: [dealers]
      summary: Create dealer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"
  /dealers/{id}:
    get:
      operationId: getDealer
      tags: [dealers]
      summary: Get dealer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dealer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"
    patch:
      operationId: updateDealer
      tags: [dealers]
      summary: Update dealer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealerUpdate"
      responses:
        "200":
          description: Updated dealer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"

  # ─── INVENTORY / WAREHOUSES ───────────────────────────────────────────────────
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [inventory]
      summary: List warehouses
      responses:
        "200":
          description: Warehouse list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [inventory]
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - in: query
          name: warehouseId
          schema:
            type: integer
        - in: query
          name: lowStock
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"

  # ─── BANNERS ─────────────────────────────────────────────────────────────────
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # ─── WARRANTY ────────────────────────────────────────────────────────────────
...
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List blogs
      responses:
        "200":
          description: Blog list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
    post:
      operationId: createBlog
      tags: [blogs]
      summary: Create blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
  /blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get blog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    patch:
      operationId: updateBlog
      tags: [blogs]
      summary: Update blog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    delete:
      operationId: deleteBlog
      tags: [blogs]
      summary: Delete blog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUPPORT TICKETS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Dealer:
      type: object
      required: [id, businessName, contactName, email, status, createdAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        discountPercent:
          type: number
        totalOrders:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        createdAt:
          type: string
    DealerList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Dealer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DealerInput:
      type: object
      required: [businessName, contactName, email]
      properties:
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        discountPercent:
          type: number
    DealerUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        discountPercent:
          type: number

    # ─── WAREHOUSE ─────────────────────────────────────────────────────────────
    Warehouse:
      type: object
      required: [id, name, city, address]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        totalItems:
          type: integer
    WarehouseInput:
      type: object
      required: [name, city, address]
      properties:
        name:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
    InventoryItem:
      type: object
      required: [id, productId, productName, warehouseName, stock, lowStockThreshold]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        warehouseName:
          type: string
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean

    # ─── BANNER ────────────────────────────────────────────────────────────────
...
    Lead:
      type: object
      required: [id, name, email, source, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        source:
          type: string
          enum: [contact_form, dealer_form, whatsapp, call_request, product_inquiry]
        status:
          type: string
          enum: [new, contacted, interested, follow_up, converted, closed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LeadInput:
      type: object
      required: [name, email, source]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        source:
          type: string
          enum: [contact_form, dealer_form, whatsapp, call_request, product_inquiry]
        notes:
          type: ["string", "null"]
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, interested, follow_up, converted, closed]
        notes:
          type: ["string", "null"]

    # ─── WARRANTY ──────────────────────────────────────────────────────────────
...
    Blog:
      type: object
      required: [id, title, slug, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
    BlogInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]

    # ─── TICKET ────────────────────────────────────────────────────────────────
...