# lib/api-zod/src/generated/types/dealerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora Enterprise Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface DealerInput {
  businessName: string;
  contactName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  city?: string | null;
  discountPercent?: number;
}